fix: rename ICEsat2_SI_tools => tools
