+++ src/icesat2_tracks/analysis_db/B03_plot_spectra_ov.py
@@ -11,8 +11,8 @@ import xarray as xr
 from matplotlib.gridspec import GridSpec
 import typer
 
-import icesat2_tracks.ICEsat2_SI_tools.iotools as io
-import icesat2_tracks.ICEsat2_SI_tools.generalized_FT as gFT
+import icesat2_tracks.tools.iotools as io
+import icesat2_tracks.tools.generalized_FT as gFT
 import icesat2_tracks.local_modules.m_tools_ph3 as MT
 from icesat2_tracks.local_modules import m_general_ph3 as M
 from icesat2_tracks.config.IceSAT2_startup import (
