+++ src/icesat2_tracks/tools/angle_optimizer.py
