+++ src/icesat2_tracks/tools/filter_regrid.py
