+++ src/icesat2_tracks/tools/read_ground_tracks.py
