+++ src/icesat2_tracks/tools/spicke_remover.py
